+++ src/main/java/shop/sajotuna/order/orders/service/OrderStatusService.java
@@ -2,7 +2,6 @@ package shop.sajotuna.order.orders.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.sajotuna.order.coupon.domain.UserCoupon;
@@ -19,15 +18,11 @@ import shop.sajotuna.order.point.service.PointService;
 import shop.sajotuna.order.point.service.dto.event.PointEvent;
 import shop.sajotuna.order.stock.service.StockService;
 
-import java.time.LocalDateTime;
-import java.util.List;
 import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
 public class OrderStatusService {
-
-    private static final String SCHEDULE = "0 0 12 * * *";
 
     private final OrderRepository orderRepository;
     private final ApplicationEventPublisher eventPublisher;
@@ -41,16 +36,6 @@ public class OrderStatusService {
     public void shippedOrder(Long orderId) {
         Order order = orderRepository.findById(orderId).orElseThrow(OrderNotFoundException::new);
         order.shipped();
-    }
-
-    // 배송중으로 변경된 주문은 일정 시간경과 후 완료 처리 됨
-    @Scheduled(cron = SCHEDULE) // 매일 낮 12시 마다 실행됨
-    @Transactional
-    public void deliveredOrder() {
-        // 현재 시간 기준으로 1일 이상 지난 주문들을 가져온다
-        List<Order> orders = orderRepository.findShippedOrders(LocalDateTime.now().minusDays(1));
-        // 배송 날짜와 1일 이상 차이가 난다면 배송완료로 변경
-        orders.forEach(Order::delivered);
     }
 
     // 주문 반품 처리
